+++ SampleApp/Finish.js
@@ -2,7 +2,7 @@ import React from 'react';
 import {StyleSheet, Text, View} from 'react-native';
 import {Link} from 'react-router-native';
 
-export default (props) => {
+export default props => {
   return (
     <View style={styles.container}>
       <Text style={styles.welcome}>Status:</Text>