+++ SampleApp/backend-server-example.js
@@ -41,8 +41,8 @@ const createSdkToken = async (applicant, applicationId) => {
 
   await applicantResponse
     .json()
-    .then((responseJson) => (sdkRequestBody.applicant_id = responseJson.id))
-    .catch((err) => {
+    .then(responseJson => (sdkRequestBody.applicant_id = responseJson.id))
+    .catch(err => {
       console.log(err, 'error');
       return {
         status: 'Unable to start the SDK',
@@ -74,8 +74,8 @@ const createSdkToken = async (applicant, applicationId) => {
   let sdkToken;
   await sdkTokenResponse
     .json()
-    .then((responseJson) => (sdkToken = responseJson.token))
-    .catch((err) => {
+    .then(responseJson => (sdkToken = responseJson.token))
+    .catch(err => {
       console.log(err, 'error');
       return {
         status: 'Unable to start the SDK',
@@ -115,8 +115,8 @@ const createSdkToken = async (applicant, applicationId) => {
 
     await workflowRunIdResponse
       .json()
-      .then((responseJson) => (workflowRunId = responseJson.id))
-      .catch((err) => {
+      .then(responseJson => (workflowRunId = responseJson.id))
+      .catch(err => {
         console.log(err, 'error');
         return {
           status: 'Unable to start the SDK',